Preview/Clang-Tidy: rewrite to match coderabbit support recipe
