+++ .clang-tidy
@@ -1,0 +1,16 @@
+Checks: >-
+  -*,
+  clang-analyzer-*,
+  bugprone-*,
+  modernize-use-nullptr,
+  modernize-use-override,
+  cppcoreguidelines-owning-memory,
+  cppcoreguidelines-pro-type-member-init
+WarningsAsErrors: >-
+  clang-analyzer-*,
+  bugprone-*,
+  modernize-use-nullptr,
+  modernize-use-override,
+  cppcoreguidelines-owning-memory,
+  cppcoreguidelines-pro-type-member-init
+FormatStyle: none
