--- clang-tidy/.clang-tidy
@@ -1,16 +1,0 @@
----
-Checks: >
-  bugprone-*,
-  cert-*,
-  clang-analyzer-*,
-  cppcoreguidelines-*,
-  hicpp-*,
-  modernize-*,
-  performance-*,
-  portability-*,
-  readability-*,
-  -modernize-use-trailing-return-type
-
-WarningsAsErrors: ''
-HeaderFilterRegex: '.*'
-FormatStyle: file
