--- clang-tidy/CMakeLists.txt
@@ -1,10 +1,0 @@
-cmake_minimum_required(VERSION 3.15)
-project(clang_tidy_demo CXX)
-
-set(CMAKE_CXX_STANDARD 17)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
-set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
-
-add_compile_options(-Wall -Wextra)
-
-add_executable(demo test.cpp)
